- refactor package-manager and import-map related things:   - in `/src/importmap/typedefs.ts`, purge the old `RuntimePackageMetadata` abstract class (which was succeeded by the new variant 3-commits prior).   - rename `/src/importmap/funcdefs.ts` to `/src/importmap/mod.ts`, since the `/src/importmap/` directory will no longer contain package management submodules.   - to organize package management submodules:     - add a new directory `/src/packageman/` dedicated to runtime package management submodules.     - move `/src/importmap/base.ts` to `/src/packageman/base.ts`.     - move `/src/importmap/deno.ts` to `/src/packageman/deno.ts`.
